Avances en proyecto de administracion de cursos

+++ Gestion-Cursos/src/main/java/gestionCursos/Docente/Docente.java
@@ -3,9 +3,11 @@ package gestionCursos.Docente;
 import gestionCursos.Curso.Curso;
 import gestionCursos.Estudiantes.Estudiante;
 
+import java.math.BigDecimal;
+
 public interface Docente {
 
     boolean estaHabilitadoParaDictarClasesA(Estudiante estudiante, Curso curso);
     boolean esProfesionalGuardavida();
-    double loQueCobraPor(Curso curso);
+    BigDecimal loQueCobraPor(Curso curso);
 }
